added acceptance test for space-magnitude binning and catalog filtering included tolerance for magnitude units on binning routine modified magnitude_bins function to reduce issues coming from floating point addition updated bin1d_vec function to account for absolute error in floating point numbers fixed ascii reader for cases where catalog_id is None removed io.py file and added those routines to __init__.py added more routines to __all__ in __init__.py
