+++ csep/core/regions.py
@@ -7,11 +7,10 @@ from xml.etree import ElementTree as ET
 # Third-party imports
 import matplotlib.path
 import numpy
-
-# PyCSEP imports
 import numpy as np
 import pyproj
 
+# PyCSEP imports
 from csep.utils.calc import bin1d_vec
 from csep.utils.scaling_relationships import WellsAndCoppersmith
 
@@ -195,13 +194,20 @@ def magnitude_bins(start_magnitude, end_magnitude, dmw):
     Returns:
         bin_edges (numpy.ndarray)
     """
-    return numpy.arange(start_magnitude, end_magnitude+dmw/2, dmw)
+    # convert to integers to prevent accumulating floating point errors
+    const = 10000
+    start = numpy.floor(const * start_magnitude)
+    end = numpy.floor(const * end_magnitude)
+    d = const * dmw
+    return (numpy.arange(start, end + d / 2, d) / const)
 
 def create_space_magnitude_region(region, magnitudes):
     """Simple wrapper to create space-magnitude region """
     if not isinstance(region, CartesianGrid2D):
         raise TypeError("region must be CartesianGrid2D")
     # bind to region class
+    if magnitudes is None:
+        raise ValueError("magnitudes should not be None if creating space-magnitude region.")
     region.magnitudes = magnitudes
     region.num_mag_bins = len(region.magnitudes)
     return region
@@ -370,7 +376,7 @@ def compute_vertices(origin_points, dh, tol=numpy.finfo(float).eps):
     """
     return list(map(lambda x: compute_vertex(x, dh, tol=tol), origin_points))
 
-def _bin_catalog_spatio_magnitude_counts(lons, lats, mags, n_poly, mask, idx_map, binx, biny, mag_bins):
+def _bin_catalog_spatio_magnitude_counts(lons, lats, mags, n_poly, mask, idx_map, binx, biny, mag_bins, tol=0.00001):
     """
     Returns a list of event counts as ndarray with shape (n_poly, n_cat) where each value
     represents the event counts within the polygon.
@@ -388,7 +394,7 @@ def _bin_catalog_spatio_magnitude_counts(lons, lats, mags, n_poly, mask, idx_map
     # index in 2d grid
     idx = bin1d_vec(lons, binx)
     idy = bin1d_vec(lats, biny)
-    mag_idxs = bin1d_vec(mags, mag_bins, right_continuous=True)
+    mag_idxs = bin1d_vec(mags, mag_bins, tol=tol, right_continuous=True)
     # start with zero event counts in each bin
     event_counts = numpy.zeros((n_poly, len(mag_bins)))
     # does not seem that we can vectorize this part
