+++ docs/conf.py
@@ -28,9 +28,9 @@ html_show_copyright = False
 html_show_sphinx = False
 
 # The short X.Y version
-version = 'v0.1'
+version = 'v0.2'
 # The full version, including alpha/beta/rc tags
-release = 'v0.1.0'
+release = 'v0.2.0'
 
 
 # -- General configuration ---------------------------------------------------
